Refactored Cli code and tests into seperate module Cleans things up a bit and seperates the concerns
