+++ core/src/main.rs
@@ -5,13 +5,13 @@
 
 pub mod base;
 mod command;
+mod cli;
 use base::{Score, Scores};
-use clap::{Parser, Subcommand};
+use crate::cli::*;
+use clap::Parser;
 use command::*;
-use home::home_dir;
 use lazy_static::lazy_static;
 use notify::{watcher, DebouncedEvent, Watcher};
-use url::Url;
 use std::{
     fs::File,
     io::{Read, Write},
@@ -21,7 +21,6 @@ use std::{
     time::Duration,
 };
 use tauri::Manager;
-use tokio::runtime::Runtime;
 
 lazy_static! {
     pub static ref ARK_SHELF_WORKING_DIR: PathBuf = PathBuf::from(Cli::parse().path);
@@ -29,81 +28,6 @@ lazy_static! {
         .join(".ark")
         .join("shelf")
         .join("scores");
-}
-
-#[derive(Parser, Default, Debug)]
-#[clap(
-    name = "ARK Shelf Desktop",
-    about = "Desktop Version of ARK Shelf, put you bookmarks when surfing."
-)]
-struct Cli {
-    #[clap(
-        short, long, help = "Path to store .link file", 
-        default_value_t = format!("{}/.ark-shelf",home_dir().expect("Can't find home dir").display())
-    )]
-    path: String,
-    #[clap(subcommand)]
-    link: Option<Link>
-}
-
-impl Cli {
-    fn add_new_link(&self) -> bool {
-        if let Some(link) = &self.link {
-            match link {
-                Link::Add(l) => {
-                    let title = l.title.clone();
-                    let desc = l.description.clone();
-                    let url = l.url.clone();
-                    create_link(title, desc, url, self.path.clone()).expect("Creating Link");
-                    return true
-                }
-            }
-        } 
-        return false
-    }
-}
-
-#[derive(Subcommand, Debug)]
-enum Link {
-    /// Adds a new link
-    Add(AddLink)
-}
-
-#[derive(Parser, Debug)]
-struct AddLink {
-    #[clap(short, long)]
-    url: String,
-
-    #[clap(short, long)]
-    title: String,
-
-    #[clap(short, long)]
-    description: Option<String>,
-}
-
-/// Creates a `.link`
-/// 
-/// Modified version of `command::create_link` which can't be reused as 
-/// there's no way to construct `tauri::State`
-fn create_link(
-    title: String,
-    desc: Option<String>,
-    url: String,
-    cli_path: String,
-) -> Result<(), String> {
-    let url = match Url::parse(url.as_str()) {
-        Ok(val) => val,
-        Err(e) => return Err(e.to_string()),
-    };
-    let resource = arklib::id::ResourceId::compute_bytes(url.as_ref().as_bytes())
-        .expect("Error compute resource from url");
-    let domain = url.domain().expect("Url has no domain");
-    let path = format!("{}/{domain}-{}.link", cli_path.clone(), resource.crc32);
-    let mut link = arklib::link::Link::new(url, title, desc);
-    let rt  = Runtime::new().map_err(|_| "Creating runtime")?;
-    let write = link.write_to_path(cli_path, path, true);
-    rt.block_on(async { write.await.expect("Writing link to path"); });
-    Ok(())
 }
 
 // Initialize file watcher.
@@ -178,7 +102,7 @@ fn init_score_watcher(path: String, scores: Arc<Mutex<Scores>>) {
 }
 
 fn main() {
-    let cli = Cli::parse();
+    let cli = crate::Cli::parse();
 
     std::fs::create_dir_all(ARK_SHELF_WORKING_DIR.as_path().join(".ark").join("shelf")).unwrap();
     lazy_static::initialize(&ARK_SHELF_WORKING_DIR);
@@ -233,21 +157,4 @@ fn main() {
         })
         .run(tauri::generate_context!())
         .expect("error while running tauri application");
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-
-    #[test]
-    fn add_link() {
-        let mut cli = Cli::default();
-        cli.path = format!("{}/.ark-shelf",home_dir().expect("Can't find home dir").display());
-        cli.link = Some(Link::Add( AddLink {
-            url: "http://example.com".into(),
-            title: "test".into(),
-            description: None
-        }));
-        cli.add_new_link();
-    }
 }
